рефактор батонов
