+++ src/components/movie-page/movie-page.jsx
@@ -4,6 +4,7 @@ import MoviesList from "../movies-list/movies-list.jsx";
 import Header from "../header/header.jsx";
 import PropTypes from "prop-types";
 import Tabs from "../tabs/tabs.jsx";
+import ControlsBtnList from "../controls-btn-list/controls-btn-list.jsx";
 
 const MoviePage = (props) => {
   const {films, currentMovie, currentGenre, onFilterClick, authorizationStatus} = props;
@@ -31,19 +32,11 @@ const MoviePage = (props) => {
           </p>
 
           <div className="movie-card__buttons">
-            <button className="btn btn--play movie-card__button" type="button">
-              <svg viewBox="0 0 19 19" width="19" height="19">
-                <use xlinkHref="#play-s"></use>
-              </svg>
-              <span>Play</span>
-            </button>
-            <button className="btn btn--list movie-card__button" type="button">
-              <svg viewBox="0 0 19 20" width="19" height="20">
-                <use xlinkHref="#add"></use>
-              </svg>
-              <span>My list</span>
-            </button>
+
+            <ControlsBtnList />
+
             <a href="add-review.html" className="btn movie-card__button">Add review</a>
+
           </div>
         </div>
       </div>
